requirements.txt
discord.py
python-dotenv
beautifulsoup4
aiohttp
flask
requests
Flask-Session
flask_bootstrap
praw==7.8.1
Flask
Flask-Limiter
gunicorn==21.2.0
psycopg2-binary
asyncpg